MB-58343 When verifying backup ignore docs not in `_default`
Before this commit when we check the result of `cbriftdump` we would
include documents in other collections. This becomes a problem on 7.6
because we added the `_system` scope. To fix we just ignore docs in any
other collection but `_default`.

Currently this works, but we may need to reassess if we add tests that
create collections.

Change-Id: I6b0e7aea42e9c7f50893be0f7b7ac243f8050d21
Reviewed-on: https://review.couchbase.org/c/examinador/+/198574
Tested-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>