MB-61658 After creating a bucket check query ready
In 7.6.2 query fixed a bug where backups would fail in mixed-mode due to
the `_system` scope not existing. They now handle this but a side-effect
is that their backup endpoint will return an error until it has checked
whether the scope exists.

For new buckets this happens within a few seconds, but in examinador we
run backups almost as soon as we have created a bucket, leading to
failures. To fix, we poll for query's backup endpoint to be ready.

NOTE: Fixup some small things the linter was complaining about in other
files.

Change-Id: I3179e6aff6e2616c25dc85dc0be32d29dcd559dc
Reviewed-on: https://review.couchbase.org/c/examinador/+/209669
Tested-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>