MB-61300 Correctly wait for flush to finish
Before this commit we detected a flush had finished by the total items
reported by `cbstats` as being 0. This is not quite right as vBuckets
are deleted asynchronously and if a vBucket had no items in it we could
think the flush is finished before it has.

To fix we instead wait for all the vBucket UUIDs to change. When a
vBucket is recreated the UUID is guaranteed to be different.

Change-Id: I56cc19247cc30a5b36f20087080f51cb78e5a648
Reviewed-on: https://review.couchbase.org/c/examinador/+/212163
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Tested-by: Matt Hall <matt.hall@couchbase.com>