MB-61300 Get item count more around flushes
We still are getting flaky failures when waiting for documents to be
persisted when loaded right after a flush. External testing to
examinador suggests we are using the right conditions to check for the
finish of a flush, pointing the finger at the data load.

We therefore get the current item count in a couple of new places. This
will be used to check that we are loading the data we think we are
loading.

Change-Id: I9faaf56145cf38c442d9e060ea450bac795a28a9
Reviewed-on: https://review.couchbase.org/c/examinador/+/216472
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Tested-by: Matt Hall <matt.hall@couchbase.com>